Moved spell checker classes to core folder
