+++ nw/core/__init__.py
@@ -3,9 +3,15 @@
 from nw.core.document import NWDoc
 from nw.core.index import NWIndex
 from nw.core.project import NWProject
+from nw.core.spellcheck import NWSpellCheck
+from nw.core.spellcheck import NWSpellEnchant
+from nw.core.spellcheck import NWSpellSimple
 
 __all__ = [
     "NWDoc",
     "NWIndex",
     "NWProject",
+    "NWSpellCheck",
+    "NWSpellEnchant",
+    "NWSpellSimple",
 ]
