+++ nw/core/spellcheck.py
