+++ nw/gui/dialogs/configeditor.py
@@ -39,7 +39,7 @@ from PyQt5.QtWidgets import (
 )
 
 from nw.additions import QSwitch, QConfigLayout
-from nw.tools import NWSpellCheck, NWSpellSimple, NWSpellEnchant
+from nw.core import NWSpellCheck, NWSpellSimple, NWSpellEnchant
 from nw.constants import nwAlert, nwQuotes
 
 logger = logging.getLogger(__name__)
