+++ nw/gui/statusbar.py
@@ -34,7 +34,7 @@ from PyQt5.QtCore import Qt, QTimer
 from PyQt5.QtGui import QColor, QPixmap, QFont
 from PyQt5.QtWidgets import QStatusBar, QLabel
 
-from nw.tools import NWSpellCheck
+from nw.core import NWSpellCheck
 
 logger = logging.getLogger(__name__)
