+++ nw/tools/__init__.py
@@ -3,9 +3,6 @@
 from nw.tools.analyse import TextAnalysis
 from nw.tools.legacy import projectMaintenance
 from nw.tools.optionstate import OptionState
-from nw.tools.spellcheck import NWSpellCheck
-from nw.tools.spellcheck import NWSpellEnchant
-from nw.tools.spellcheck import NWSpellSimple
 from nw.tools.translate import numberToWord
 from nw.tools.wordcount import countWords
 
@@ -13,9 +10,6 @@ __all__ = [
     "TextAnalysis",
     "projectMaintenance",
     "OptionState",
-    "NWSpellCheck",
-    "NWSpellEnchant",
-    "NWSpellSimple",
     "numberToWord",
     "countWords",
 ]
